macOS: Cleanup and re-org
