+++ MacElsewhen/Model/MenuTag.swift
@@ -1,0 +1,12 @@
+//
+//  MenuTag.swift
+//  MenuTag
+//
+//  Created by David Stephens on 17/09/2021.
+//
+
+enum MenuTag: Int {
+    case about
+    case services
+    case preferences
+}
